requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-admin-material-dashboard==1.0.10
django-filter==24.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
multidict==6.0.5
phonenumbers==8.13.35
phonenumberslite==8.13.35
pillow==10.3.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.5.0
twilio==9.0.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
